% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bim_to_granges}
\alias{bim_to_granges}
\title{Read a PLINK .bim file as GRanges}
\usage{
bim_to_granges(bim_file)
}
\arguments{
\item{bim_file}{Path to a .bim file.}
}
\value{
A GRanges object with SNP positions (1-based, width 1).
}
\description{
Converts a PLINK .bim file to a GRanges object with hg19 coordinates.
This is a convenience function for users who have .bim files from
1000 Genomes Phase 3.
}
\examples{
# Create a minimal mock .bim file
tmp <- tempfile(fileext = ".bim")
bim <- data.frame(1, c("rs1", "rs2"), 0,
    c(1000, 2000), "A", "G")
write.table(bim, tmp, sep = "\t",
    row.names = FALSE, col.names = FALSE, quote = FALSE)
gr <- bim_to_granges(tmp)
gr

}
