% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{CATaN-accessors}
\alias{CATaN-accessors}
\alias{tfsampleLoading}
\alias{trasampleLoading}
\alias{ccParameters}
\alias{snpScores}
\alias{topSnps}
\alias{bottomSnps}
\alias{tfsampleLoading,CATaNResult-method}
\alias{trasampleLoading,CATaNResult-method}
\alias{ccParameters,CATaNResult-method}
\alias{snpScores,CATaNAnnotation-method}
\alias{topSnps,CATaNAnnotation-method}
\alias{bottomSnps,CATaNAnnotation-method}
\title{Accessor generics for CATaN classes}
\usage{
tfsampleLoading(object)

trasampleLoading(object)

ccParameters(object)

snpScores(object)

topSnps(object)

bottomSnps(object)

\S4method{tfsampleLoading}{CATaNResult}(object)

\S4method{trasampleLoading}{CATaNResult}(object)

\S4method{ccParameters}{CATaNResult}(object)

\S4method{snpScores}{CATaNAnnotation}(object)

\S4method{topSnps}{CATaNAnnotation}(object)

\S4method{bottomSnps}{CATaNAnnotation}(object)
}
\arguments{
\item{object}{A CATaNResult or CATaNAnnotation object.}
}
\value{
A DataFrame or list depending on the accessor.
}
\description{
Generic accessor functions for CATaNResult and
CATaNAnnotation objects.
}
\examples{
set.seed(42)
tf <- matrix(sample(0:1, 500, replace = TRUE), 50, 10)
rownames(tf) <- paste0("g", seq_len(50))
colnames(tf) <- paste0("TF", seq_len(10))
ex <- matrix(rnorm(250), 50, 5)
rownames(ex) <- rownames(tf)
colnames(ex) <- paste0("s", seq_len(5))
res <- run_cca(tf, ex, n_cc = 3L, n_hvg = 30L,
               verbose = FALSE)
tfsampleLoading(res)
trasampleLoading(res)
ccParameters(res)
}
