% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CATaNResult-class}
\alias{CATaNResult-class}
\alias{show,CATaNResult-method}
\title{CATaNResult class}
\usage{
\S4method{show}{CATaNResult}(object)
}
\value{
Prints a summary of the object to the console
(invisibly returns NULL).
}
\description{
S4 class to store results from CCA of TF-GRN and
transcriptome matrices. This is the output of \code{\link{run_cca}} and
serves as input to \code{\link{align_cc_to_snps}}.
}
\section{Slots}{

\describe{
\item{\code{tf_matrix}}{Filtered TF-GRN matrix (gene x TF) used in CCA.}

\item{\code{transcriptome}}{Filtered and normalised logCPM matrix (gene x sample)
used in CCA.}

\item{\code{hvg}}{Character vector of highly variable gene names selected for CCA.}

\item{\code{svd_result}}{List containing the raw SVD output (\code{u}, \code{d},
\code{v}).}

\item{\code{n_cc}}{Integer, the number of canonical components retained.}

\item{\code{tf_sample_loading}}{DataFrame of TF sample loadings (u matrix), TF x CC.}

\item{\code{transcriptome_sample_loading}}{DataFrame of transcriptome sample loadings (v matrix), sample x CC.}

\item{\code{parameters}}{DataFrame of CCA summary parameters including variance
explained, singular value proportions, SCF, and canonical correlations.}

\item{\code{sample_metadata}}{DataFrame of sample annotations (optional).}

\item{\code{call}}{The matched call used to produce this object.}
}}

\examples{
res <- new("CATaNResult")
show(res)
}
