% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_tf_matrix}
\alias{download_tf_matrix}
\title{Download TF-gene binding matrix from Zenodo}
\usage{
download_tf_matrix(dest_dir, url = NULL, cache = TRUE, verbose = TRUE)
}
\arguments{
\item{dest_dir}{Character. Directory to save the file. Created if it does
not exist.}

\item{url}{Character. URL of the gzipped matrix file. If NULL (default),
uses the default Zenodo URL.}

\item{cache}{Logical. Whether to use BiocFileCache (default TRUE).}

\item{verbose}{Logical (default TRUE).}
}
\value{
Character. Path to the downloaded (uncompressed) matrix file.
}
\description{
Downloads the TF-gene binding association matrix used in the CATaN paper.
}
\examples{
\donttest{
matrix_file <- download_tf_matrix("~/catan_data")
tf_mat <- read.table(matrix_file, header = TRUE, row.names = 1)
}

}
